Refactor env_lock to a shared helper and use it for the h2 tests
